[I25-240] feat: 캐러셀 이미지 로딩 방식 변경 및 API 제거
